Minor refactoring of automated tests
Pull our some common functions from automated tests in preparation for
writing additional tests that all need these functions.
